Fixes #370: moves ovirt to internal directory
