--- generate.go
@@ -1,5 +1,0 @@
-package main
-
-//go:generate go run github.com/hashicorp/terraform-plugin-docs/cmd/tfplugindocs
-
-//go:generate go run scripts/get_test_image/get_test_image.go
