+++ internal/ovirt/data_source_ovirt_blank_template.go
