+++ internal/ovirt/resource_affinity_group.go
