+++ internal/ovirt/resource_ovirt_disk.go
