+++ internal/ovirt/resource_ovirt_disk_from_image.go
