+++ internal/ovirt/resource_ovirt_template.go
