+++ internal/ovirt/resource_ovirt_vm_graphics_consoles.go
