+++ internal/ovirt/testimage
@@ -1,0 +1,1 @@
+../../testimage
