--- ovirt/testimage
@@ -1,1 +1,0 @@
-../testimage
